Consolidated Themed Icons
- Added themed AntDesign Icons
