--- src/Shared/IconTheme/IoniconsThemed.tsx
@@ -1,9 +1,0 @@
-import { useTheme } from "@react-navigation/native";
-import { Ionicons } from "@expo/vector-icons"
-
-import { IoniconsThemedProps } from "./types"
-
-export function IoniconsThemed(props: IoniconsThemedProps) {
-    const { colors } = useTheme();
-    return <Ionicons {...props} color={props.text === "secondary" ? colors.secondaryText : colors.text} />
-}
