+++ src/Shared/IconTheme/types.ts
@@ -1,8 +1,10 @@
-import { FontAwesome, Ionicons } from "@expo/vector-icons";
+import { AntDesign, FontAwesome, Ionicons } from "@expo/vector-icons";
 import { ComponentProps } from "react";
 
 
 type TextTypes = "primary" | "secondary";
+type Text = { text?: TextTypes }
 
-export type IoniconsThemedProps = ComponentProps<typeof Ionicons> & { text?: TextTypes };
-export type FontAwesomeThemedProps = ComponentProps<typeof FontAwesome> & { text?: TextTypes };
+export type IoniconsThemedProps = ComponentProps<typeof Ionicons> & Text
+export type FontAwesomeThemedProps = ComponentProps<typeof FontAwesome> & Text
+export type AntDesignThemedProps = ComponentProps<typeof AntDesign> & Text
